Update ESLint Configs
The base eslint config contains most, if not all, of the config settings we need. However, we need to specify the "root" property so eslint knows not to look further up in the directory structure for more config settings.

Eslint configs were added to all package roots as well, with only the necessary updates to the base config. This ensures that all packages as close to the same linter settings as possible.