Workaround Frontend Postinstall on Windows
Yarn should be able to resolve electron-builder from the frontend's package.json, but this is broken on windows.

To workaround this issue, the postinstall script in the frontend has been renamed "afterinstall" so it will no longer be called in the normal install script flow. Instead, a postinstall script is added to the root package that calls this script. Since the install has completed, the electron-builder binary has been added and can be invoked without any errors.