Upgrade Dependencies
This commit updates the repo by upgrading dependencies using `yarn upgrade-interactive`. These versions still fall within the same constraints of the dependency list, but they are newer then the previous versions stored in the yarn.lock file.

The eslint version has been pinned to 5.12.0 to work with the latest version of react-scripts. react-scripts pins a specific version to ensure compatibility.