Update Repo Configuration (#410)
* Consolidate Dev Dependencies to Root

This commit moves any shared/common dev dependencies to the root package.json and out of the individual packages so there is only one reference version.

* Update ESLint Configs

The base eslint config contains most, if not all, of the config settings we need. However, we need to specify the "root" property so eslint knows not to look further up in the directory structure for more config settings.

Eslint configs were added to all package roots as well, with only the necessary updates to the base config. This ensures that all packages as close to the same linter settings as possible.

* Update Babel Configuration

This commit updates the babel configuration to make sure all subpackages are using the root babel config. This required adding the --root-mode upward flag when invoking babel as well as changing the config file name to be "babel.config.js"

This commit also adds the --no-comments flag when building relevant subpackages since the built babel files don't require comments to function properly. This decreases the size of files built by babel.

* Use Yarn instead of NPM

This commit updates the package configurations to use yarn for all scripts instead of using npm. This keeps the repo consistent on using one package manager.

* Workaround Frontend Postinstall on Windows

Yarn should be able to resolve electron-builder from the frontend's package.json, but this is broken on windows.

To workaround this issue, the postinstall script in the frontend has been renamed "afterinstall" so it will no longer be called in the normal install script flow. Instead, a postinstall script is added to the root package that calls this script. Since the install has completed, the electron-builder binary has been added and can be invoked without any errors.

* Use Latest Yarn on Travis (Workaround)

This commit updates the travis ci script to install the latest yarn version until a later version is installed by default.

* Upgrade Dependencies

This commit updates the repo by upgrading dependencies using `yarn upgrade-interactive`. These versions still fall within the same constraints of the dependency list, but they are newer then the previous versions stored in the yarn.lock file.

The eslint version has been pinned to 5.12.0 to work with the latest version of react-scripts. react-scripts pins a specific version to ensure compatibility.