Fix Lints
This commit fixes linter errors reported after running the `lint` script. The errors were trivial to fix.

The `worker_threads` module is not able to be resolved by eslint even though it does exist with the current version of node being used. This may be due to the experimental status of the module, but the linter error for that has been disabled for now until a more permanent fix is found.