App Performance Improvements (#406)
* mitigate logging + mitigate redundant task status emissions

* added size retrieval on frontend + more optimizations

* added removing tasks from log when app shuts down

* switched to discord.js webhook

* clean dependencies

* Temporarily commented out logging

* batch update

* batch status update changes

* action/reducer changes for batch updates

* batching status updates done

* Apply suggestions from code review

* Fix Lints

This commit fixes linter errors reported after running the `lint` script. The errors were trivial to fix.

The `worker_threads` module is not able to be resolved by eslint even though it does exist with the current version of node being used. This may be due to the experimental status of the module, but the linter error for that has been disabled for now until a more permanent fix is found.

* Fix Failing Tests

This commit fixes failing unit tests due to changes in this branch. The following changes are made:
- app.test.jsx - Update format of task event to match new format
- app.test.jsx - Pass initial task list in mock store to prevent undefined error
- taskActions.test.js - Update stop tasks to ensure that task is running (a new change)
- taskActions.test.js - Update status task input to match new format
- taskActions.test.js - removed unnecessary test related to validation (now handled through the reducer)
- taskListReducer.test.js - Update status task action to match new format
- taskReducer.test.js - Update helper functions to allow custom expected return values. Use addition to fix tests related to edit size tests (chosenSizes needed to update as well)
- taskReducer.test.js - Fix wrong helper function being called in existing task tests
- logTaskRow.test.jsx - Update props to match expected prop type format. Output is now a string instead of an object
- logTaskRow.jsx - Use destructuring to remove redundant "task" prefixes

* Remove Worker Threads References

This commit removes the exported references of the SplitWorkerThreadTaskManager and the usage of it in the frontend until #412 is addressed

* Add Babel Runtime to Project

This commit adds the babel runtime to the project and adds a transform to the babel config to allow the runtime to be used correctly. This allows the code to work properly when run.