Release Beta 7.2 (#415)
* Prevent double windows when handling auth event (#407)

This commit updates logic surrounding activating and deactivating to ensure that only one main window and one auth window are open at a time.

* Update Travis Node Version (#408)

* Update Travis Node Version

This commit updates the travis config to match the packaged version of node for electron 4. This should help make sure we are syncing our testing environment to what will be run on users' machines.

* Update Travis Cache Settings

This commit updates the travis cache settings to the recommended setting for yarn.

* Update Repo Configuration (#410)

* Consolidate Dev Dependencies to Root

This commit moves any shared/common dev dependencies to the root package.json and out of the individual packages so there is only one reference version.

* Update ESLint Configs

The base eslint config contains most, if not all, of the config settings we need. However, we need to specify the "root" property so eslint knows not to look further up in the directory structure for more config settings.

Eslint configs were added to all package roots as well, with only the necessary updates to the base config. This ensures that all packages as close to the same linter settings as possible.

* Update Babel Configuration

This commit updates the babel configuration to make sure all subpackages are using the root babel config. This required adding the --root-mode upward flag when invoking babel as well as changing the config file name to be "babel.config.js"

This commit also adds the --no-comments flag when building relevant subpackages since the built babel files don't require comments to function properly. This decreases the size of files built by babel.

* Use Yarn instead of NPM

This commit updates the package configurations to use yarn for all scripts instead of using npm. This keeps the repo consistent on using one package manager.

* Workaround Frontend Postinstall on Windows

Yarn should be able to resolve electron-builder from the frontend's package.json, but this is broken on windows.

To workaround this issue, the postinstall script in the frontend has been renamed "afterinstall" so it will no longer be called in the normal install script flow. Instead, a postinstall script is added to the root package that calls this script. Since the install has completed, the electron-builder binary has been added and can be invoked without any errors.

* Use Latest Yarn on Travis (Workaround)

This commit updates the travis ci script to install the latest yarn version until a later version is installed by default.

* Upgrade Dependencies

This commit updates the repo by upgrading dependencies using `yarn upgrade-interactive`. These versions still fall within the same constraints of the dependency list, but they are newer then the previous versions stored in the yarn.lock file.

The eslint version has been pinned to 5.12.0 to work with the latest version of react-scripts. react-scripts pins a specific version to ensure compatibility.

* App Performance Improvements (#406)

* mitigate logging + mitigate redundant task status emissions

* added size retrieval on frontend + more optimizations

* added removing tasks from log when app shuts down

* switched to discord.js webhook

* clean dependencies

* Temporarily commented out logging

* batch update

* batch status update changes

* action/reducer changes for batch updates

* batching status updates done

* Apply suggestions from code review

* Fix Lints

This commit fixes linter errors reported after running the `lint` script. The errors were trivial to fix.

The `worker_threads` module is not able to be resolved by eslint even though it does exist with the current version of node being used. This may be due to the experimental status of the module, but the linter error for that has been disabled for now until a more permanent fix is found.

* Fix Failing Tests

This commit fixes failing unit tests due to changes in this branch. The following changes are made:
- app.test.jsx - Update format of task event to match new format
- app.test.jsx - Pass initial task list in mock store to prevent undefined error
- taskActions.test.js - Update stop tasks to ensure that task is running (a new change)
- taskActions.test.js - Update status task input to match new format
- taskActions.test.js - removed unnecessary test related to validation (now handled through the reducer)
- taskListReducer.test.js - Update status task action to match new format
- taskReducer.test.js - Update helper functions to allow custom expected return values. Use addition to fix tests related to edit size tests (chosenSizes needed to update as well)
- taskReducer.test.js - Fix wrong helper function being called in existing task tests
- logTaskRow.test.jsx - Update props to match expected prop type format. Output is now a string instead of an object
- logTaskRow.jsx - Use destructuring to remove redundant "task" prefixes

* Remove Worker Threads References

This commit removes the exported references of the SplitWorkerThreadTaskManager and the usage of it in the frontend until #412 is addressed

* Add Babel Runtime to Project

This commit adds the babel runtime to the project and adds a transform to the babel config to allow the runtime to be used correctly. This allows the code to work properly when run.

* Task Log Expansion (#411)

* mitigate logging + mitigate redundant task status emissions

* added size retrieval on frontend + more optimizations

* added removing tasks from log when app shuts down

* small changes

* switched to discord.js webhook

* Temporarily commented out logging

* batch update

* batch status update changes

* pr changes

* Apply suggestions from code review

* revert package.json changes && travis.yml updates

* task log expansion started

* log expansion header and row sizing adjusted

* added proxy event sender

* product name returned and shown now

* live log changes + additions

* style changes

* clear log selected and focused on stop task

* cleanup + mass link/password base functionality

* task runner side of mass overrides

* revert task runner side of things

* cleanup windows.js file

* css changes

* fix linter problems

* small change, fix lint

* progress on swapping proxy

* temp change dev logger level + css changes to live log

* added auto-scrolling to live log

* some pr changes

* proxy swapping updates/pr changes

* rm worker threads stuff..

* fix restock mode proxy issue

* some test fixes

* fix lint

* added migration tests

* added bring the heat bots

* added loopback for OOS/payment failure

* logTask test coverage

* rm unnecessary mount in test

* pr changes

* rm moment dependency

* rm unused chalk dependency

* i'm a big dumb

* test changes

* bump version

* replace yarn lock

* menu bug fixed + some small changes to about/auth page

* window.Bridge.close not working in about window.. not sure why

* Filter Out Tasks With Invalid "index" Property (#416)

* added test

* renamed test describe

* change id to index

* Update packages/frontend/src/state/migrators/v0.4.0/index.js

Co-Authored-By: walmat <matthew.wallt@gmail.com>

* rm unnecessary state changes

* fix test

* Added temporary stoppage for special parsers in restock mode

* added Travisscott subdomains for May 3rd