Add Obfuscation to Build Pipeline
This commit modifies the task-runner package.json to add obfucation to the build pipeline.

The build script is updated to run two sub commands: build:babel and build:obfuscate. The main export file of the package is changed to point to the obfuscated code instead of the built code.

Obfuscation options are added to a separate config file so the cli command remains clean.