Add Task Runner Built package
This commit adds a new packages that will contain the built, obfuscated code of the task-runner. This package functions exactly the same as the task-runner package, but is safe to include in built packages since no unobfuscated source is contained in it.

This package is meant purely as a proxy and no source code nor development should occur here. The task-runner package is updated to change the obfuscated output to this package.

The frontend can then depend on this package so the packed app will contain only obfuscated task-runner code.