Obfuscate lib folder
This commit updates the frontend package to obfuscate the lib folder. The javascript obfuscator is used for all files in the lib folder and the obfuscated code is moved to a dist/ folder.

The main file is updated to use the obfuscated code and the build script is updated to obfuscate code as well as bundle the react side.

NOTE: When running in dev mode, the unobfuscated version is used.