Sync Captcha Window Themes to Main Window (#435)
* Store Theme options in Captcha Window Manager

This commit updates the Captcha Window Manager to store the background color received from a create captcha window request. This is one of the ways the react side notifies the electron side of the captcha window theme, so adding this keep the states in sync

* Use Captcha Theme Options

This commit updates the createCaptchaWindow call in the Captcha Window Manager to use the stored captcha theme options. This allows the stored captcha window background color to always be used.

* Handle Change Theme Requests

This commit updates the WindowManager to handle ChangeTheme requests from the react side. For now, only the captcha windows will support theme changes, so this function defers implementation to the captcha window manager.

The captcha window manager updates the stored theme options. Future commits will handle changing the theme for running windows.

* Send Change Theme Event to Open Captcha Windows

This commit updates the Captcha Window Manager to send an event to change the theme of the window. This will allow windows to respond to changes appropriately.

* Spread options instead of setting them

This commit updates the change theme handler in the Captcha Window Manager to spread the incoming theme options instead of setting them. This allows the incoming options to be merged correctly into the existing options.

* Respond to Change Theme Requests in Captcha Window

This commit updates the Captcha preload to add methods for registering/deregistering callbacks to a theme change.

The captcha.js file is updated to use the preload methods, and adjusts the document.body style with the incoming background color.

* Send initial theme options on App mount

This commit updates the App component to notify the electron side of the starting theme on component mount. This allows the theme options to sync initially instead of waiting for a setTheme call or captcha window launch through the react side.

tests are updated to pass (including a mock function in the mocked bridge) and tests are added to ensure the initial call is made.

* bug fix