Fix some failing tests
This commit fixes failing tests due to the migrator referencing the same `rates` array in the initialState. This is an issue when testing because the initial state is used, but not an issue in production because new references are created for every action.

Instead of using the same reference, a new reference is creating using spread operator to get the existing elements of the array. This prevents the errors from occurring in the tests, but gives the same result in production.