Add DSML Default Rates (#434)
* started migrator for DSML rates

* added all rates in constants file

* progress on migrator

* added in shipping manager profile

* changing countries will now change the dsm rate properly

* update current profile to reset to no rates. that way no rates transfer over

* rm changes

* remove update adding rates

* added check to change country when updating

* Switch to using test ids

* Fix some failing tests

This commit fixes failing tests due to the migrator referencing the same `rates` array in the initialState. This is an issue when testing because the initial state is used, but not an issue in production because new references are created for every action.

Instead of using the same reference, a new reference is creating using spread operator to get the existing elements of the array. This prevents the errors from occurring in the tests, but gives the same result in production.

* Handle Update Existing Rates Edge Case

This commit updates the profileListReducer to handle an edge case where a pre-fetched dsml rate existed, but the new updated country does not support pre-fetched rates.

The old existing rate is deleted in this case so we don't use the now invalid pre-fetched rate.