Start Adding Shared Monitor Implementation
This commit adds the beginnings of a Shared Monitor that can handle multiple monitoring for multiple product inputs (all on the same site). The implementation is similar to the TaskRunner, but includes different events that are handled.

All shared/easy methods are implemented, while other methods are stubbed. These stubbed methods are subject to change during development as new ideas on implementation come up.