Update Match to allow Configurable Input and Type
This commit updates the singular match function in the parser to allow a match input and type to be injected in. The current implementation still functions as expected.

This commit deprecates the singular task input and parse type in favor of multiple task inputs to parse. Updating the match function in a non-breaking way allows us to move toward that goal while keeping the functionality of the rest of the parsers intact.