Add Fetch Method
This commit updates the regular parsers to use a new fetch method that fetches the product data. This method is broken out of the run method so it can be used directly instead of within the run method alone.

The fetch method is implemented in each variant and used in the run method to prevent breaking the current implementation.