Refactor MatchKeywords to work with multiple keyword pairs
This commit updates the matchKeywords implementation to work with multiple pairs. Instead of a single match result getting returned with a single keyword pair as an input, an array is returned that is ordered the same for the input array of keywords.

Keyword matching is also updated to use case-insensitive regexes instead of the indexOf string method. This removes the need to convert everything to uppercase -- a slight optimization

This commit also updates the usages around the task-runner project to conform to the new api. This should provide backwards compatibility while we transition to using shared monitors.