Rename ParseType to ProductInputType
This commit updates the naming of the ParseType enumeration to ProductInputType to be more accurate of what the enumeration represents.

The enumeration declaration is moved to the constants file so it can live among other constant declarations. All variable names and functions are renamed to match the change accordingly.

No functionality is changed in this commit.