Add ParserType Enum
This commit adds a ParserType enumeration that describes the type of parser used to fetch product data. This can be used to identify the type of parser used through the new `.type` getter of a Parser instance.

The ParserType can also be used to identify what type of parser was used to capture product data for a single product object. Since the Atom and Xml do not contain full product info, this is a way to tag a product data object and tell whether or not more data is needed.

As of now, this feature is not used, but the tags are added to product data so they can be used in the future.