Initial Implementation of Parse all method
This commit provides an initial implementation of the parse all method based on the steps described within the comments of the method.

The parse all method will be split up into the handler methods to perform the tasks concurrently, but this requires a setup of class level data structures that hasn't been decided yet. This single method implementation should provide a picture into what data structures are necessary and what data structures can be moved to a class-level in order to complete the split into handler methods.