Redact Html Inputs Before Logging
This commit updates sensitive inputs by tagging them with the data-private tag to prevent LogRocket from rendering the input. This ensures that no personal/sensitive data is sent to LogRocket.

This commit successfully adds redactions for all inputs, but there are still a couple of select components that aren't redacted. This is likely due to the use of react-select for the dropdown select components and more research will have to done to integrate the two.